Enable command php katana post (#8)
In the archive Katana.php was inserted into registerCommands method,
one call for the PostCommand Class that set parameter of the command
inserted through terminal.

The PostCommand Class does one call of the PostBuilder class and send
feedback for user. In PostBuilder the name file for new Post is defined
through method nameFile, where the date, the slug of the title and the
blade extension or markdown extension are concatenated. the file is save
on /content/_blog